moved build to build_config
